154/ convert temperature scales

+++ app/src/main/java/is/hello/sense/units/UnitConverter.java
@@ -9,5 +9,5 @@ public interface UnitConverter {
      */
     UnitConverter IDENTITY = v -> v;
 
-    double convert(float value);
+    Float convert(Float value);
 }
